Move all of switch overrde service to switch.py
