+++ homeassistant/components/geniushub/__init__.py
@@ -73,17 +73,6 @@ ATTR_DURATION = "duration"
 
 SVC_SET_ZONE_MODE = "set_zone_mode"
 SVC_SET_ZONE_OVERRIDE = "set_zone_override"
-SVC_SET_SWITCH_OVERRIDE = "set_switch_override"
-
-SET_SWITCH_OVERRIDE_SCHEMA = vol.Schema(
-    {
-        vol.Required(ATTR_ENTITY_ID): cv.entity_id,
-        vol.Optional(ATTR_DURATION): vol.All(
-            cv.time_period,
-            vol.Range(min=timedelta(minutes=5), max=timedelta(days=1)),
-        ),
-    }
-)
 
 SET_ZONE_MODE_SCHEMA = vol.Schema(
     {
@@ -162,33 +151,6 @@ def setup_service_functions(hass: HomeAssistantType, broker):
         }
 
         async_dispatcher_send(hass, DOMAIN, payload)
-
-    async def set_switch_override(call) -> None:
-        """Set the system mode."""
-        entity_id = call.data[ATTR_ENTITY_ID]
-
-        registry = await hass.helpers.entity_registry.async_get_registry()
-        registry_entry = registry.async_get(entity_id)
-
-        if registry_entry is None or registry_entry.platform != DOMAIN:
-            raise ValueError(f"'{entity_id}' is not a known {DOMAIN} entity")
-
-        if registry_entry.domain != "switch":
-            raise ValueError(f"'{entity_id}' is not an {DOMAIN} zone")
-
-        payload = {
-            "unique_id": registry_entry.unique_id,
-            "service": call.service,
-            "data": call.data,
-        }
-        async_dispatcher_send(hass, DOMAIN, payload)
-
-    hass.services.async_register(
-        DOMAIN,
-        SVC_SET_SWITCH_OVERRIDE,
-        set_switch_override,
-        schema=SET_SWITCH_OVERRIDE_SCHEMA,
-    )
 
     hass.services.async_register(
         DOMAIN, SVC_SET_ZONE_MODE, set_zone_mode, schema=SET_ZONE_MODE_SCHEMA
