+++ homeassistant/components/geniushub/switch.py
@@ -2,12 +2,31 @@
 from datetime import timedelta
 from typing import Optional
 
+import voluptuous as vol
+
 from homeassistant.components.switch import DEVICE_CLASS_OUTLET, SwitchEntity
+from homeassistant.const import ATTR_ENTITY_ID
+from homeassistant.core import callback
+from homeassistant.helpers import config_validation as cv
+from homeassistant.helpers.dispatcher import async_dispatcher_send
+from homeassistant.helpers.service import verify_domain_control
 from homeassistant.helpers.typing import ConfigType, HomeAssistantType
 
-from . import ATTR_DURATION, DOMAIN, SVC_SET_SWITCH_OVERRIDE, GeniusZone
+from . import ATTR_DURATION, DOMAIN, GeniusZone
 
 GH_ON_OFF_ZONE = "on / off"
+
+SVC_SET_SWITCH_OVERRIDE = "set_switch_override"
+
+SET_SWITCH_OVERRIDE_SCHEMA = vol.Schema(
+    {
+        vol.Required(ATTR_ENTITY_ID): cv.entity_id,
+        vol.Optional(ATTR_DURATION): vol.All(
+            cv.time_period,
+            vol.Range(min=timedelta(minutes=5), max=timedelta(days=1)),
+        ),
+    }
+)
 
 
 async def async_setup_platform(
@@ -25,6 +44,41 @@ async def async_setup_platform(
             for z in broker.client.zone_objs
             if z.data["type"] == GH_ON_OFF_ZONE
         ]
+    )
+
+    setup_service_functions(hass, broker)
+
+
+@callback
+def setup_service_functions(hass: HomeAssistantType, broker):
+    """Set up the service functions."""
+
+    @verify_domain_control(hass, DOMAIN)
+    async def set_switch_override(call) -> None:
+        """Set the system mode."""
+        entity_id = call.data[ATTR_ENTITY_ID]
+
+        registry = await hass.helpers.entity_registry.async_get_registry()
+        registry_entry = registry.async_get(entity_id)
+
+        if registry_entry is None or registry_entry.platform != DOMAIN:
+            raise ValueError(f"'{entity_id}' is not a known {DOMAIN} entity")
+
+        if registry_entry.domain != "switch":
+            raise ValueError(f"'{entity_id}' is not an {DOMAIN} zone")
+
+        payload = {
+            "unique_id": registry_entry.unique_id,
+            "service": call.service,
+            "data": call.data,
+        }
+        async_dispatcher_send(hass, DOMAIN, payload)
+
+    hass.services.async_register(
+        DOMAIN,
+        SVC_SET_SWITCH_OVERRIDE,
+        set_switch_override,
+        schema=SET_SWITCH_OVERRIDE_SCHEMA,
     )
